Move schema update tests to schema package
